+++ autotsk/lib/add_task/add_task_db.dart
@@ -11,6 +11,8 @@ class AddFormMethod {
     required String uid,
     required String title,
     required String date,
+    String? startTime,
+    String? endTime,
     required String priority,
     required String location,
     required String notes,
@@ -24,6 +26,8 @@ class AddFormMethod {
           'uid': uid,
           'title': title,
           'date': date,
+          'start time' : startTime,
+          'end time': endTime, 
           'priority': priority,
           'location': location,
           'notes': notes,
