+++ autotsk/pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   get:
   get_storage:
   quickalert: ^1.0.1
+  sqflite:
 
 dev_dependencies:
   flutter_lints: ^2.0.0
